Extract the HTTP request router to lib/http-router.js
Verbatim move of the http.createServer callback as a named handler
(handleHttpRequest) plus BINARY_FILE_TYPES, whose single user is the
router's binary file transport. Mechanical conversions, each
count-asserted at extraction: 6 __dirname sites become ROOT_DIR (one
hop up from lib/, same repo or app.asar root the code always read),
3 workspace resolves read getWorkspace() at use time, and the
root-owned capabilities go through wired deps: isInsideWorkspace,
getFileTreeCached, safeSend, the chatProcesses and
pendingPermissionRequests maps as identity accessors, and the
permission timeout as a getter. Direct lib requires, no wiring, same
names: discoverAgents, boundaryGrantCovers, recordEvent,
resolvePermissionConvoId.

The root keeps the server itself: http.createServer(handler), the
WebSocket server bootstrap, and the connection state that lived below
the router (chatProcesses, the auto-resume circuit breaker,
connectedClients, disconnectBuffer) stay where they were.

Zero test retargets: the HTTP behaviour pins drive a booted server and
keep exercising the moved code through the wired module.

server.js: 5,404 -> 5,188 lines. Suite 1462 green.
